refactor: 로또 기본 가격 long 타입으로 변경

+++ src/main/java/lotto/constant/LottoConstant.java
@@ -4,7 +4,7 @@ public final class LottoConstant {
     public final static int LOTTO_NUMBER_COUNT = 6;
     public final static int MIN_LOTTO_NUMBER = 1;
     public final static int MAX_LOTTO_NUMBER = 45;
-    public final static int BASIC_LOTTO_PRICE = 1000;
+    public final static long BASIC_LOTTO_PRICE = 1000;
     public final static int PERCENT = 100;
 
     private LottoConstant() {
